django_rockitchef/rockitchef_project/requirements.txt
Django==2.1.7
django-taggit==0.24.0
django-taggit-serializer==0.1.7
djangorestframework==3.9.2
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytz==2018.9
six==1.12.0
beautifulsoup4==4.6.0
fake-useragent==0.1.11
Jinja2==2.10
jsondiff==1.1.2
PyYAML==3.12
pysqlite==2.8.3
requests==2.21.0
virtualenv==16.4.3
